# common/api/swagger.json
{
  "openapi": "3.0.0",
  "components": {
...
    "schemas": {
      "FieldErrors": {
        "properties": {},
        "type": "object",
        "additionalProperties": {
          "properties": {
            "value": {},
            "message": {
              "type": "string"
            }
          },
          "required": ["message"],
          "type": "object"
        }
      },
      "Pick_ValidateError.status-or-message-or-fields_": {
        "properties": {
          "status": {
            "type": "number",
            "format": "double"
          },
          "message": {
            "type": "string"
          },
          "fields": {
            "$ref": "#/components/schemas/FieldErrors"
          }
        },
        "required": ["status", "message", "fields"],
        "type": "object",
        "description": "From T, pick a set of properties whose keys are in the union K"
      },
      "Pick_Error.message_": {
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": ["message"],
        "type": "object",
        "description": "From T, pick a set of properties whose keys are in the union K"
      }
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "name": "access_token",
        "in": "query"
      }
    }
  },
  "info": {
    "title": "api",
    "version": "0.1.0",
    "contact": {}
  },
  "paths": {
...
    "/profiles": {
      "get": {
        "operationId": "GetAllEvents",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Example 1": {
                    "value": {
                      "result": [
                        {
                          "id": "ffca2a3a-c0bf-4931-b65a-8d8ccf867096",
                          "firstName": "John",
                          "lastName": "Doe",
                          "academicTitle": "Dr.",
                          "email": "",
                          "url": null,
                          "avatar": "https://img.platform.org/public/url/of/profile/avatar",
                          "background": "https://img.platform.org/public/url/of/profile/background",
                          "bio": "I am a </strong>bio</strong>",
                          "areas": [
                            {
                              "area": {
                                "name": "India"
                              }
                            },
                            {
                              "area": {
                                "name": "Bavaria"
                              }
                            }
                          ],
                          "offers": [
                            {
                              "offer": {
                                "title": "Offer 1"
                              }
                            },
                            {
                              "offer": {
                                "title": "Offer 2"
                              }
                            }
                          ],
                          "seekings": [
                            {
                              "offer": {
                                "title": "Offer 1"
                              }
                            },
                            {
                              "offer": {
                                "title": "Offer 2"
                              }
                            }
                          ]
                        }
                      ],
                      "skip": 0,
                      "take": 2
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "A take parameter larger than 50 is not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "status": {
                      "type": "number",
                      "enum": [400],
                      "nullable": false
                    },
                    "message": {
                      "type": "string",
                      "enum": ["Parameter out of range"],
                      "nullable": false
                    },
                    "fields": {
                      "properties": {
                        "take": {
                          "properties": {
                            "message": {
                              "type": "string",
                              "enum": [
                                "A take parameter larger than 50 is not allowed"
                              ],
                              "nullable": false
                            }
                          },
                          "required": ["message"],
                          "type": "object"
                        }
                      },
                      "required": ["take"],
                      "type": "object"
                    }
                  },
                  "required": ["status", "message", "fields"],
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pick_ValidateError.status-or-message-or-fields_"
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "status": 401,
                      "message": "Authentication failed",
                      "fields": {
                        "access_token": {
                          "message": "Invalid access token"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pick_ValidateError.status-or-message-or-fields_"
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "status": 422,
                      "message": "Validation failed",
                      "fields": {
                        "skip": {
                          "message": "'skip' is required"
                        },
                        "take": {
                          "message": "'take' is required"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Pick_Error.message_"
                    },
                    {
                      "properties": {
                        "status": {
                          "type": "number",
                          "format": "double"
                        }
                      },
                      "required": ["status"],
                      "type": "object"
                    }
                  ]
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "status": 500,
                      "message": "Internal Server Error"
                    }
                  }
                }
              }
            }
          }
        },
        "description": "Retrieves all profiles of the community including their public information.\nThey do not include private (Community-only) fields.",
        "summary": "Retrieve all profiles.",
        "tags": ["Profiles"],
        "security": [
          {
            "api_key": []
          }
        ],
        "parameters": [
          {
            "description": "The number of items to skip before starting to collect the result set",
            "in": "query",
            "name": "skip",
            "required": true,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "description": "The number of items to return (A take parameter larger than 50 is not allowed)",
            "in": "query",
            "name": "take",
            "required": true,
            "schema": {
              "format": "double",
              "type": "number"
            }
          }
        ]
      }
    },
...
    "/events": {
      "get": {
        "operationId": "GetAllEvents",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Example 1": {
                    "value": {
                      "result": [
                        {
                          "id": "a24004fa-aee8-40e2-aeb8-c6a40a1a2ee4",
                          "name": "Consensus in heterogeneous groups",
                          "slug": "event-slug",
                          "url": "https://community.platform.org/url/to/event",
                          "background": "https://img.platform.org/public/url/of/event/background",
                          "description": "Welcome to our event where we present and discuss multiple ways to find consensus in heterogeneous groups. In <strong>multiple workshops</strong> and keynote speeches we will explore strategies to ...",
                          "subline": "Discussing different approaches to find consensus",
                          "startTime": "2017-06-17T10:00:00.000Z",
                          "endTime": "2017-06-18T18:00:00.000Z",
                          "participationFrom": "2017-06-12T00:00:00.000Z",
                          "participationUntil": "2017-06-16T00:00:00.000Z",
                          "participantLimit": 20,
                          "venueName": "Meetingpalace",
                          "venueStreet": "Gatheringstreet",
                          "venueStreetNumber": "12",
                          "venueCity": "City of Happenings",
                          "venueZipCode": "12345",
                          "canceled": false,
                          "parentEventId": "some-parent-event-id",
                          "areas": [
                            {
                              "area": {
                                "name": "India"
                              }
                            },
                            {
                              "area": {
                                "name": "Bavaria"
                              }
                            },
                            {
                              "area": {
                                "name": "New York City"
                              }
                            }
                          ],
                          "types": [
                            {
                              "eventType": {
                                "title": "Exchange"
                              }
                            },
                            {
                              "eventType": {
                                "title": "Keynote"
                              }
                            },
                            {
                              "eventType": {
                                "title": "Workshop"
                              }
                            }
                          ],
                          "focuses": [
                            {
                              "focus": {
                                "title": "Social Media Management"
                              }
                            },
                            {
                              "focus": {
                                "title": "Politics"
                              }
                            },
                            {
                              "focus": {
                                "title": "Science"
                              }
                            },
                            {
                              "focus": {
                                "title": "Career Orientation"
                              }
                            }
                          ],
                          "tags": [
                            {
                              "tag": {
                                "title": "Didactics"
                              }
                            },
                            {
                              "tag": {
                                "title": "Gender"
                              }
                            },
                            {
                              "tag": {
                                "title": "Innovation"
                              }
                            },
                            {
                              "tag": {
                                "title": "Networking"
                              }
                            }
                          ],
                          "targetGroups": [
                            {
                              "targetGroup": {
                                "title": "Elementary School"
                              }
                            },
                            {
                              "targetGroup": {
                                "title": "Early Childhood Education"
                              }
                            },
                            {
                              "targetGroup": {
                                "title": "Teachers"
                              }
                            }
                          ],
                          "experienceLevel": {
                            "title": "Beginner"
                          },
                          "stage": {
                            "title": "Hybrid"
                          },
                          "_count": {
                            "participants": 20,
                            "waitingList": 3
                          },
                          "responsibleOrganizations": [
                            {
                              "organization": {
                                "name": "Platform",
                                "slug": "platform",
                                "logo": "https://img.platform.org/public/url/of/logo",
                                "url": "https://community.platform.org/url/to/organization"
                              }
                            }
                          ]
                        },
                        {
                          "id": "ffca2a3a-c0bf-4931-b65a-8d8ccf867096",
                          "name": "Smallest Event",
                          "slug": "smallest-event",
                          "url": null,
                          "background": null,
                          "description": null,
                          "subline": null,
                          "startTime": "2017-06-17T10:00:00.000Z",
                          "endTime": "2017-06-18T18:00:00.000Z",
                          "participationFrom": "2017-06-12T00:00:00.000Z",
                          "participationUntil": "2017-06-16T00:00:00.000Z",
                          "participantLimit": null,
                          "venueName": null,
                          "venueStreet": null,
                          "venueStreetNumber": null,
                          "venueCity": null,
                          "venueZipCode": null,
                          "canceled": true,
                          "parentEventId": null,
                          "areas": [],
                          "types": [],
                          "focuses": [],
                          "tags": [],
                          "targetGroups": [],
                          "experienceLevel": null,
                          "stage": null,
                          "_count": {
                            "participants": 0,
                            "waitingList": 0
                          },
                          "responsibleOrganizations": []
                        }
                      ],
                      "skip": 0,
                      "take": 2
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "A take parameter larger than 50 is not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "status": {
                      "type": "number",
                      "enum": [400],
                      "nullable": false
                    },
                    "message": {
                      "type": "string",
                      "enum": ["Parameter out of range"],
                      "nullable": false
                    },
                    "fields": {
                      "properties": {
                        "take": {
                          "properties": {
                            "message": {
                              "type": "string",
                              "enum": [
                                "A take parameter larger than 50 is not allowed"
                              ],
                              "nullable": false
                            }
                          },
                          "required": ["message"],
                          "type": "object"
                        }
                      },
                      "required": ["take"],
                      "type": "object"
                    }
                  },
                  "required": ["status", "message", "fields"],
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pick_ValidateError.status-or-message-or-fields_"
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "status": 401,
                      "message": "Authentication failed",
                      "fields": {
                        "access_token": {
                          "message": "Invalid access token"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pick_ValidateError.status-or-message-or-fields_"
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "status": 422,
                      "message": "Validation failed",
                      "fields": {
                        "skip": {
                          "message": "'skip' is required"
                        },
                        "take": {
                          "message": "'take' is required"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Pick_Error.message_"
                    },
                    {
                      "properties": {
                        "status": {
                          "type": "number",
                          "format": "double"
                        }
                      },
                      "required": ["status"],
                      "type": "object"
                    }
                  ]
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "status": 500,
                      "message": "Internal Server Error"
                    }
                  }
                }
              }
            }
          }
        },
        "description": "Retrieves all events of the community including their public information.\nThey do not include private (Community-only) fields.",
        "summary": "Retrieve all events.",
        "tags": ["Events"],
        "security": [
          {
            "api_key": []
          }
        ],
        "parameters": [
          {
            "description": "The number of items to skip before starting to collect the result set",
            "in": "query",
            "name": "skip",
            "required": true,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "description": "The number of items to return (A take parameter larger than 50 is not allowed)",
            "in": "query",
            "name": "take",
            "required": true,
            "schema": {
              "format": "double",
              "type": "number"
            }
          }
        ]
      }
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}